+++ src/randp.c
@@ -175,7 +175,7 @@ randp_bytes(void* buf, size_t n) [[gnu::nonnull]]
 
         uint8_t* src = &this_->pool[RANDP_NUM_BYTES - this_->rand_bytes_remaining];
 
-        const size_t m = MIN(n, (size_t)this_->rand_bytes_remaining);
+        const int m = (int)MIN(n, (size_t)this_->rand_bytes_remaining);
 
         (void)memcpy(dst, src, m);
         explicit_bzero(src, m);
